DataIngestor/requirements.txt
kafka-python>=2.0.2
influxdb-client>=1.36.0
